crlf to lf fix php 7.4 -> 7.2 fix run app fix run front remove db with legacy refactoring
