--- .env
@@ -1,1 +1,0 @@
-POSTGRES_PASSWORD=12345
