+++ app/.env
@@ -1,5 +1,1 @@
-DB_NAME=postgres
-DB_USER=postgres
-DB_PASSWORD=12345
-DB_HOST=db
-DB_TYPE=pgsql
+TOKEN_ACCU='mf6zSCSKxUVAD46e9a6xZMk5UY2lq7Xf'
