+++ app/entrypoint.sh
@@ -1,3 +1,2 @@
 composer install
-app migrations:migrate -n --allow-no-migration
 php-fpm
