+++ app/src/HTTP/City.php
@@ -2,7 +2,7 @@
 
 namespace App\HTTP;
 
-use App\WheatherAPI\Contracts\BridgeInterface;
+use App\WheatherInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -10,22 +10,17 @@ use Zend\Diactoros\Response\JsonResponse;
 
 class City implements RequestHandlerInterface
 {
-
-
     /**
-     * @var BridgeInterface
+     * @var WheatherInterface
      */
-    private BridgeInterface $api;
+    private $wheather;
 
-    private array $cities = [
-        'tashkent',
-        'moscow',
-        'london',
-    ];
+    private $cities;
 
-    public function __construct(BridgeInterface $api)
+    public function __construct(WheatherInterface $api, array $citiesAvailable)
     {
-        $this->api = $api;
+        $this->wheather = $api;
+        $this->cities = $citiesAvailable;
     }
 
     public function handle(ServerRequestInterface $request): ResponseInterface
@@ -34,7 +29,7 @@ class City implements RequestHandlerInterface
         if (!in_array($city, $this->cities)) {
             return new JsonResponse(['City not Found'], 404, ['Access-Control-Allow-Origin' => '*']);
         }
-        $weather = $this->api->getWeather($city);
+        $weather = $this->wheather->getWeather($city);
         return new JsonResponse($weather, 200, ['Access-Control-Allow-Origin' => '*']);
     }
 }
