--- app/src/WheatherAPI/Contracts/BridgeInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\WheatherAPI\Contracts;
-
-interface BridgeInterface
-{
-    public function getWeather(string $city);
-}
