+++ app/src/WheatherInterface.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace App;
+
+interface WheatherInterface
+{
+    public function getWeather(string $city);
+}
