+++ docker-compose.yml
@@ -14,10 +14,6 @@ services:
       target: developer
     volumes:
       - ./app:/app
-    environment:
-      DB_PASSWORD: ${POSTGRES_PASSWORD}
-#    depends_on:
-#      - db
   front:
     build:
       context: front
@@ -28,9 +24,3 @@ services:
       - app
     ports:
       - 8081:8080
-#  db:
-#    image: postgres
-#    environment:
-#      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-#    ports:
-#      - 5432:5432
