+++ front/cmd.sh
@@ -1,2 +1,3 @@
+
 npm install
 npm run serve
